validations in bank service

+++ src/main/java/common/bankarskiSistem/service/BankService.java
@@ -1,5 +1,6 @@
 package common.bankarskiSistem.service;
 
+import common.bankarskiSistem.exceptions.NameOfTheBankAlreadyExistException;
 import common.bankarskiSistem.model.Bank;
 import common.bankarskiSistem.model.BankAccount;
 import common.bankarskiSistem.model.ExchangeRates;
@@ -20,10 +21,23 @@ public class BankService {
 
     /**
      *
-     * @param bank object of Bank
+     * @param name of the bank
+     * @param address of the bank
      * @return bank that was created
+     * @throws NameOfTheBankAlreadyExistException
      */
-    public Bank createBank(Bank bank) {
+    public Bank createBank(String name, String address) throws NameOfTheBankAlreadyExistException {
+        if (name == null)
+            throw new NullPointerException("The name is null.");
+        if (address == null)
+            throw new NullPointerException("The address is null.");
+
+        Bank bank = new Bank();
+        bank.setName(name);
+        bank.setAddress(address);
+
+        if(bankRepository.findByName(bank.getName()).isPresent())
+            throw new NameOfTheBankAlreadyExistException("Name of the bank already exists.");
         return bankRepository.save(bank);
     }
 
@@ -32,6 +46,8 @@ public class BankService {
      * @param idBank of the bank to be deleted
      */
     public void deleteBank(Integer idBank) {
+        if(bankRepository.findById(idBank).isPresent())
+            throw new NullPointerException("The bank does not exist.");
         bankRepository.deleteById(idBank);
     }
 
@@ -41,6 +57,10 @@ public class BankService {
      * @param idBank of the bank
      */
     public void updateBankName(String name, Integer idBank) {
+        if (name == null)
+            throw new NullPointerException("The name is null.");
+        if(bankRepository.findById(idBank).isPresent())
+            throw new NullPointerException("The bank does not exist.");
         bankRepository.getReferenceById(idBank).setName(name);
     }
 
@@ -50,6 +70,10 @@ public class BankService {
      * @param idBank od the bank
      */
     public void updateBankAddress(String address, Integer idBank) {
+        if (address == null)
+            throw new NullPointerException("The address is null.");
+        if(bankRepository.findById(idBank).isPresent())
+            throw new NullPointerException("The bank does not exist.");
         bankRepository.getReferenceById(idBank).setAddress(address);
     }
 
@@ -59,6 +83,8 @@ public class BankService {
      * @return exchange rate that was created
      */
     public ExchangeRates createExchangeRates(ExchangeRates exchangeRates) {
+        if(exchangeRates == null)
+            throw new NullPointerException("The excenge rate is null.");
         return exchangeRatesRepository.save(exchangeRates);
     }
 
@@ -67,6 +93,8 @@ public class BankService {
      * @return exchange rates
      */
     public Optional<ExchangeRates> getExcangeRates(Bank bank) {
+        if (bank == null)
+            throw new NullPointerException("The bank is null.");
         return bankRepository.getExchangeRates(bank);
 
     }
@@ -83,7 +111,11 @@ public class BankService {
      * @return list of users
      */
     public List<User> getAllUsers(Bank bank) {
+        if (bank == null)
+            throw new NullPointerException("The bank is null.");
         List<User> users = new ArrayList<>();
+        if (bankRepository.getBankAccounts(bank) == null)
+            throw new NullPointerException("There is no accounts in this bank.");
         for (BankAccount account: bankRepository.getBankAccounts(bank)) {
             users.add(account.getUser());
         }
